fix bigs

+++ ex01/ft_strncpy.c
@@ -6,14 +6,14 @@
 /*   By: edi-iori <edi-iori@student.42lyon.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/07/07 10:51:17 by edi-iori          #+#    #+#             */
-/*   Updated: 2023/07/07 12:09:02 by edi-iori         ###   ########lyon.fr   */
+/*   Updated: 2023/07/08 20:27:53 by edi-iori         ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
 // #include <stdio.h>
 char	*ft_strncpy(char *dest, char *src, unsigned int n)
 {
-	size_t	i;
+	unsigned int	i;
 
 	i = 0;
 	while (i < n && src[i] != '\0')
